refactor: remove blockchain service classes and replace with centralized transaction execution
