--- frontend/src/services/Blockchain/IBlockchainService.ts
@@ -1,11 +1,0 @@
-import { PublicKey } from '@trustwallet/wallet-core/dist/src/wallet-core';
-
-import { tss } from '../../../wailsjs/go/models';
-
-export interface IBlockchainService {
-  executeTransaction(
-    publicKey: PublicKey,
-    txInputData: Uint8Array,
-    signatures: { [key: string]: tss.KeysignResponse }
-  ): Promise<string>;
-}
