--- frontend/src/services/Blockchain/ton/BlockchainServiceTon.ts
@@ -1,68 +1,0 @@
-import { TW } from '@trustwallet/wallet-core';
-import { PublicKey } from '@trustwallet/wallet-core/dist/src/wallet-core';
-
-import { tss } from '../../../../wailsjs/go/models';
-import { Post } from '../../../../wailsjs/go/utils/GoHttp';
-import { assertSignature } from '../../../chain/utils/assertSignature';
-import { assertErrorMessage } from '../../../lib/utils/error/assertErrorMessage';
-import { Endpoint } from '../../Endpoint';
-import { BlockchainService } from '../BlockchainService';
-import { IBlockchainService } from '../IBlockchainService';
-import SignatureProvider from '../signature-provider';
-
-export class BlockchainServiceTon
-  extends BlockchainService
-  implements IBlockchainService
-{
-  public async executeTransaction(
-    publicKey: PublicKey,
-    txInputData: Uint8Array,
-    signatures: { [key: string]: tss.KeysignResponse }
-  ): Promise<string> {
-    const publicKeyData = publicKey.data();
-
-    const preHashes = this.walletCore.TransactionCompiler.preImageHashes(
-      this.coinType,
-      txInputData
-    );
-
-    const { data, errorMessage } =
-      TW.TxCompiler.Proto.PreSigningOutput.decode(preHashes);
-
-    assertErrorMessage(errorMessage);
-
-    const allSignatures = this.walletCore.DataVector.create();
-    const publicKeys = this.walletCore.DataVector.create();
-    const signatureProvider = new SignatureProvider(
-      this.walletCore,
-      signatures
-    );
-    const signature = signatureProvider.getSignature(data);
-
-    assertSignature({
-      publicKey,
-      message: data,
-      signature,
-    });
-
-    allSignatures.add(signature);
-    publicKeys.add(publicKeyData);
-
-    const compiled = this.walletCore.TransactionCompiler.compileWithSignatures(
-      this.coinType,
-      txInputData,
-      allSignatures,
-      publicKeys
-    );
-
-    const output = TW.TheOpenNetwork.Proto.SigningOutput.decode(compiled);
-
-    assertErrorMessage(output.errorMessage);
-
-    const response = await Post(Endpoint.broadcastTonTransaction(), {
-      boc: output.encoded,
-    });
-
-    return Buffer.from(response.result.hash, 'base64').toString('hex');
-  }
-}
